refactor: 장르 관리 유틸 분리
